Create z472c678_Hw07.cpp
Went in and talked to the teacher and was given advice to use a pair vector. When I tried it out myself I got a abort code message because vector could not be classified tried to use a pair array to no avail. Some tips to getting the counter to pair store inside of an array would be appreciated.